Activity Log: Continue fetching multi-page results
It has been possible that we have more results in a request to
the Activity Log than would fit in the requests number of events
for a single page. Previously we have been ignoring everything
past the first page.

In this patch we're looking at the API results and if there are
more pages of results we will start a loop to fetch those in
sequence by passing the `page` parameter for what should be the
next page.